refactor: extract AppItem and SimpleCommand to separate files
Extracted `AppItem` and `SimpleCommand` helper classes from `MainWindow.xaml.cs` into their own files `AppItem.cs` and `SimpleCommand.cs` to improve maintainability and organization.
Updated `AGENTS.md` to reflect the new location of the `AppItem` class.
