+++ MainWindow.xaml.cs
@@ -467,20 +467,4 @@ namespace Launchbox
         }
     }
 
-    // --- HELPER CLASSES ---
-    public class AppItem
-    {
-        public string Name { get; set; } = string.Empty;
-        public string Path { get; set; } = string.Empty;
-        public BitmapImage? Icon { get; set; }
-    }
-
-    public class SimpleCommand : System.Windows.Input.ICommand
-    {
-        private readonly Action _action;
-        public SimpleCommand(Action action) => _action = action;
-        public bool CanExecute(object? parameter) => true;
-        public void Execute(object? parameter) => _action();
-        public event EventHandler? CanExecuteChanged { add { } remove { } }
-    }
 }
